chore: relocate liblinphone backend behind integration package
