+++ src/yoyopod/cli/remote/ops/commands.py
@@ -94,8 +94,8 @@ def build_native_shim_refresh_command(deploy_config: PiDeployConfig) -> str:
         "    ),\n"
         "    (\n"
         "        'Liblinphone',\n"
-        "        Path('src/yoyopod/communication/integrations/liblinphone_binding/native/build/libyoyopod_liblinphone_shim.so'),\n"
-        "        [Path('src/yoyopod/communication/integrations/liblinphone_binding/native')],\n"
+        "        Path('src/yoyopod/communication/integrations/liblinphone/native/build/libyoyopod_liblinphone_shim.so'),\n"
+        "        [Path('src/yoyopod/communication/integrations/liblinphone/native')],\n"
         "        ['uv', 'run', 'yoyoctl', 'build', 'liblinphone'],\n"
         "    ),\n"
         "]\n"
