+++ src/yoyopod/communication/integrations/liblinphone/__init__.py
@@ -1,0 +1,15 @@
+"""Liblinphone integration package for native binding and Python backend."""
+
+from yoyopod.communication.integrations.liblinphone.binding import (
+    LiblinphoneBinding,
+    LiblinphoneBindingError,
+    LiblinphoneNativeEvent,
+)
+from yoyopod.communication.integrations.liblinphone.backend import LiblinphoneBackend
+
+__all__ = [
+    "LiblinphoneBackend",
+    "LiblinphoneBinding",
+    "LiblinphoneBindingError",
+    "LiblinphoneNativeEvent",
+]
