--- src/yoyopod/communication/integrations/liblinphone_binding/__init__.py
@@ -1,13 +1,0 @@
-"""CPython binding layer for the native Liblinphone shim."""
-
-from yoyopod.communication.integrations.liblinphone_binding.binding import (
-    LiblinphoneBinding,
-    LiblinphoneBindingError,
-    LiblinphoneNativeEvent,
-)
-
-__all__ = [
-    "LiblinphoneBinding",
-    "LiblinphoneBindingError",
-    "LiblinphoneNativeEvent",
-]
